+++ ikii-back-commons/src/main/java/mx/ikii/commons/domain/OrderStatus.java
@@ -11,7 +11,8 @@ public enum OrderStatus {
 	ACCEPTED("ACCEPTED"),
 	DECLINE("DECLINE"),
 	FINISHED("FINISHED"),
-	ERROR("ERROR");
+	ERROR("ERROR"),
+	REFUNDED("REFUNDED");
 
     private final String status;
 }