+++ ikii-back-commons/src/main/java/mx/ikii/commons/payload/dto/TaxLineDTO.java
@@ -1,0 +1,12 @@
+package mx.ikii.commons.payload.dto;
+
+import java.math.BigDecimal;
+
+import lombok.Data;
+
+@Data
+public class TaxLineDTO {
+
+	private BigDecimal amount;
+	private String description;
+}