+++ ikii-back-commons/src/main/java/mx/ikii/commons/persistence/collection/OrderDetail.java
@@ -6,10 +6,16 @@ import org.bson.types.ObjectId;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 import mx.ikii.commons.persistence.collection.util.ProductDetail;
 
 @Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
 @Document(collection = "OrderDetail")
 public class OrderDetail {