+++ ikii-back-commons/src/main/java/mx/ikii/commons/persistence/collection/PaymentOrder.java
@@ -8,15 +8,23 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.DBRef;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
+import lombok.NoArgsConstructor;
+import mx.ikii.commons.payload.dto.PaymentMethodDTO;
 
 @Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
 @Document(collection = "PaymentOrder")
 public class PaymentOrder {
 	
 	@Id
 	private String id;
 	private ObjectId customerId;
+	private String customerIdConekta;
 	private String customerProviderId;
 	private String transactionId;
 	private BigDecimal total;
@@ -25,9 +33,11 @@ public class PaymentOrder {
 	private BigDecimal internalComission;
 	private BigDecimal tax;
 	private BigDecimal amountRefunded;
-	private Integer paymentMethod;
 	private String status;
 	private LocalDateTime createdAt;
+	private String message;
+	private PaymentMethodDTO paymentMethod;
+	private String reasonRefund;
 	
 	@DBRef
 	private OrderDetail detail;