+++ ikii-back-microservices/ikii-back-payment/src/main/java/mx/ikii/payment/payload/dto/LineItemsDTO.java
@@ -1,6 +1,5 @@
 package mx.ikii.payment.payload.dto;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -22,7 +21,7 @@ public class LineItemsDTO {
 	
 	private String description;
 	
-	private BigDecimal unit_price;
+	private Integer unit_price;
 	
 	private Integer quantity;