+++ ikii-back-microservices/ikii-back-payment/src/main/java/mx/ikii/payment/payload/dto/TaxLineDTO.java
@@ -1,0 +1,11 @@
+package mx.ikii.payment.payload.dto;
+
+import lombok.Data;
+
+@Data
+public class TaxLineDTO {
+
+	private Integer amount;
+	private String description;
+	
+}